Split water_dashboard.rs (607 lines) into water_dashboard/ subdirectory
Split crates/ui/src/water_dashboard.rs into focused modules:
- types.rs: WaterDashboardVisible resource, MGD_TO_GPD constant, SourceAggregation
- panels.rs: Individual UI panel rendering functions
- ui_system.rs: Main water_dashboard_ui system
- tests.rs: All unit tests
- mod.rs: Thin facade with re-exports

All public items re-exported from mod.rs — no external import changes needed.
